+++ Frends.AzureDataLake.DeleteDirectory/CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
 
 ## [1.0.0] - 2024-05-14
-### Changed
-- Initial implementation
+### Added
+- Initial implementation of Frends.AzureDataLake.DeleteDirectory.